fix directinput axises.

+++ io/input.cpp
@@ -128,7 +128,7 @@ input::~input()
 		if(bl)bl->process( di->read() );
 	}
 
-	bool input::getbutton(int which, int & value, int & retro_id, bool isanalog)
+	bool input::getbutton(int which, int16_t & value, int & retro_id, bool isanalog)
 	{
 		return bl->getbutton(which, value,retro_id,isanalog);
 	}
