+++ io/input.h
@@ -44,7 +44,7 @@ public:
 
 	// input_i_dinput
 	void poll();
-	bool getbutton(int which, int & value,int & retro_id, bool isanalog=false);
+	bool getbutton(int which, int16_t & value,int & retro_id, bool isanalog=false);
 
 	unsigned read();
